+++ src/main/java/daepoid/stockManager/api/IngredientApiController.java
@@ -57,6 +57,25 @@ public class IngredientApiController {
         return new Result(IngredientDTOs);
     }
 
+    @GetMapping("/v3/ingredients")
+    @ApiOperation(value="전체 재료 조회", notes="레시피에 할당된 재료들을 일괄적으로 조회")
+    public Result ingredientsV3(@RequestParam("firstResult") Integer firstResult,
+                                @RequestParam("maxResult") Integer maxResult) {
+        List<Ingredient> ingredients;
+
+        if(firstResult == null) {
+            ingredients = ingredientService.findIngredients(maxResult);
+        } else {
+            ingredients = ingredientService.findIngredients(firstResult, maxResult);
+        }
+
+        //엔티티 -> DTO 변환
+        List<IngredientDTO> IngredientDTOs = ingredients.stream()
+                .map(IngredientDTO::new)
+                .collect(Collectors.toList());
+        return new Result(IngredientDTOs);
+    }
+
     @PostMapping("/v1/ingredients")
     @ApiOperation(value="재료 생성", notes="재고 정보를 받아와 레시피에 들어가는 재료의 정보를 생성하고 재료의 아이디를 반환")
     public CreateIngredientResponseDTO createIngredientV1(@RequestBody @Valid CreateIngredientRequestDTO requestDTO) {
@@ -80,13 +99,22 @@ public class IngredientApiController {
     @GetMapping("/v1/ingredients/{ingredientId}")
     @ApiOperation(value="재료 조회", notes="재료 아이디를 이용하여 재료의 정보를 반환")
     public IngredientDTO findIngredientV1(@PathVariable("ingredientId") Long ingredientId) {
-        return new IngredientDTO(ingredientService.findIngredient(ingredientId));
+        Ingredient ingredient = ingredientService.findIngredient(ingredientId);
+        if(ingredient == null) {
+            throw new IllegalArgumentException("잘못된 아이디");
+        }
+        return new IngredientDTO(ingredient);
     }
 
     @PutMapping("/v1/ingredients/{ingredientId}")
     @ApiOperation(value="재료 정보 수정 - PUT", notes="재료 아이디를 이용하여 재료 정보 수정")
     public UpdateIngredientResponseDTO updateIngredientV1(@PathVariable("ingredientId") Long ingredientId,
                                                           @RequestBody @Valid UpdateIngredientRequestDTO requestDTO) {
+        Ingredient ingredient = ingredientService.findIngredient(ingredientId);
+        if(ingredient == null) {
+            throw new IllegalArgumentException("잘못된 아이디");
+        }
+
         ingredientService.changeQuantity(ingredientId, requestDTO.getQuantity());
         ingredientService.changeUnitType(ingredientId, requestDTO.getUnitType());
         ingredientService.changeUnitPrice(ingredientId, requestDTO.getUnitPrice());
@@ -99,6 +127,11 @@ public class IngredientApiController {
     @ApiOperation(value="재료 정보 수정 - PATCH", notes="재료 아이디를 이용하여 재료 정보 수정")
     public UpdateIngredientResponseDTO updatePatchIngredientV1(@PathVariable("ingredientId") Long ingredientId,
                                                          @RequestBody @Valid UpdatePatchIngredientRequestDTO requestDTO) {
+        Ingredient ingredient = ingredientService.findIngredient(ingredientId);
+        if(ingredient == null) {
+            throw new IllegalArgumentException("잘못된 아이디");
+        }
+
         if(requestDTO.getQuantity() != null) {
             ingredientService.changeQuantity(ingredientId, requestDTO.getQuantity());
         }
@@ -119,6 +152,10 @@ public class IngredientApiController {
     @ApiOperation(value="재료 정보 삭제", notes="재료 아이디를 이용하여 재료 정보 삭제")
     public DeleteIngredientResponseDTO deleteIngredientV1(@PathVariable("ingredientId") Long ingredientId) {
         Ingredient ingredient = ingredientService.findIngredient(ingredientId);
+        if(ingredient == null) {
+            throw new IllegalArgumentException("잘못된 아이디");
+        }
+
         ingredientService.deleteIngredient(ingredientId);
         return new DeleteIngredientResponseDTO(ingredient.getId());
     }
