+++ src/main/java/daepoid/stockManager/api/ItemApiController.java
@@ -53,6 +53,25 @@ public class ItemApiController {
         return new Result(ItemDTOs);
     }
 
+    @GetMapping("/v3/items")
+    @ApiOperation(value="전체 재고 조회", notes="전체 재고 정보를 조회하고 반환")
+    public Result findItemsV3(@RequestParam("firstResult") Integer firstResult,
+                              @RequestParam("maxResult") Integer maxResult) {
+        List<Item> items;
+
+        if(firstResult == null) {
+            items = itemService.findItems(maxResult);
+        } else {
+            items = itemService.findItems(firstResult, maxResult);
+        }
+
+        //엔티티 -> DTO 변환
+        List<ItemDTO> ItemDTOs = items.stream()
+                .map(ItemDTO::new)
+                .collect(Collectors.toList());
+        return new Result(ItemDTOs);
+    }
+
     @PostMapping("/v1/items")
     @ApiOperation(value="재고 품목 생성", notes="새로운 재고 품목을 생성하고 아이디를 반환")
     public CreateItemResponseDTO createItemV1(@RequestBody @Valid CreateItemRequestDTO requestDTO) {
@@ -73,13 +92,23 @@ public class ItemApiController {
     @GetMapping("/v1/items/{itemId}")
     @ApiOperation(value="재고 조회", notes="아이디를 이용하여 재고 품목의 정보를 조회")
     public ItemDTO findItemV1(@PathVariable("itemId") Long itemId) {
-        return new ItemDTO(itemService.findItem(itemId));
+        Item item = itemService.findItem(itemId);
+        if(item == null) {
+            throw new IllegalArgumentException("잘못된 아이디");
+        }
+
+        return new ItemDTO(item);
     }
 
     @PutMapping("/v1/items/{itemId}")
     @ApiOperation(value="재고 수정", notes="재고의 정보를 수정하고 수정된 정보를 반환")
     public UpdateItemResponseDTO updateItemV1(@PathVariable("itemId") Long itemId,
                                               @RequestBody @Valid UpdateItemRequestDTO requestDTO) {
+        Item item = itemService.findItem(itemId);
+        if(item == null) {
+            throw new IllegalArgumentException("잘못된 아이디");
+        }
+
         itemService.changeName(itemId, requestDTO.getName());
         itemService.changeItemType(itemId, requestDTO.getItemType());
         itemService.changePrice(itemId, requestDTO.getPrice());
@@ -94,6 +123,11 @@ public class ItemApiController {
     @ApiOperation(value="재고 수정", notes="재고의 정보를 수정하고 수정된 정보를 반환")
     public UpdateItemResponseDTO updatePatchItemV1(@PathVariable("itemId") Long itemId,
                                                    @RequestBody @Valid UpdatePatchItemRequestDTO requestDTO) {
+        Item item = itemService.findItem(itemId);
+        if(item == null) {
+            throw new IllegalArgumentException("잘못된 아이디");
+        }
+
         if(requestDTO.getName() != null) {
             itemService.changeName(itemId, requestDTO.getName());
         }
@@ -120,6 +154,10 @@ public class ItemApiController {
     @ApiOperation(value="재고 품목 삭제", notes="재고 품목 자체를 삭제하고 아이디를 반환")
     public DeleteItemResponseDTO deleteItemV1(@PathVariable("itemId") Long itemId) {
         Item item = itemService.findItem(itemId);
+        if(item == null) {
+            throw new IllegalArgumentException("잘못된 아이디");
+        }
+
         itemService.removeItem(item);
         return new DeleteItemResponseDTO(item.getId());
     }
